Mon, Sep 16, 2024, 12:20 PM +02:00

+++ indoor.py
@@ -2,8 +2,8 @@ x = input("How much was the meal? ")
 
 x = x.removeprefix("$")
 
-z = int(x)
+z = float(x)
 
-y = float(z)
+y = int(z)
 
 print(y)
